:construction: use serde to take default value
NOTE: Each struct in archive policy must have a default implementation for all their fields. And please add `#[serde(default)]` to each field. It's because that the policy may update its fields in the future, and it's hard to update all this column in the database for all outdated games.